1. add simplehttpserver example 2. change echoserver example to be echoserver
